+++ Dark Basic Public Shared/Dark Basic Pro SDK/Shared/Sound/CSoundC.cpp
@@ -640,14 +640,14 @@ DARKSDK void LoadRawSound ( LPSTR szFilename, int iID, bool bFlag, int iSilentFa
 	// Uses actual or virtual file..
 	char VirtualFilename[_MAX_PATH];
 	strcpy(VirtualFilename, szFilename);
-	g_pGlob->UpdateFilenameFromVirtualTable( (DWORD)VirtualFilename);
+	g_pGlob->UpdateFilenameFromVirtualTable( VirtualFilename);
 
 	CheckForWorkshopFile (VirtualFilename);
 
 	// Decrypt and use media, re-encrypt
-	g_pGlob->Decrypt( (DWORD)VirtualFilename );
+	g_pGlob->Decrypt( VirtualFilename );
 	LoadRawSoundCore ( VirtualFilename, iID, bFlag, iSilentFail, iGlobalSound );
-	g_pGlob->Encrypt( (DWORD)VirtualFilename );
+	g_pGlob->Encrypt( VirtualFilename );
 }
 
 DARKSDK void LoadSound ( LPSTR szFilename, int iID )
